+++ src/tko/cmds/cmd_run.py
@@ -324,15 +324,15 @@ class Run:
         pieces = target_path.split(os.sep)
         if len(pieces) >= 3:
             task.set_key(pieces[-1])
-            task.set_source_alias(pieces[-2])
+            task.set_alias(pieces[-2])
             # task.set_rep_folder(os.sep.join(pieces[:-2]))
         elif len(pieces) == 2:
             task.set_key(pieces[-1])
-            task.set_source_alias(pieces[-2])
+            task.set_alias(pieces[-2])
             # task.set_rep_folder(os.sep)
         else:
             task.set_key(pieces[-1])
-            task.set_source_alias("")
+            task.set_alias("")
             # task.set_rep_folder(os.sep)
 
         self.__task = task