+++ src/tko/util/decoder.py
@@ -1,26 +1,18 @@
 # ignore missing import 
-from typing import Any
-import chardet
 import os
 
 class Decoder:
 
     @staticmethod
-    def load(file_path: str) -> str:
-        if not os.path.exists(file_path):
 
-            raise Exception("Arquivo {} não encontrado".format(os.path.abspath(file_path)))
-        with open(file_path, "rb") as file:
-            raw_data = file.read()
-            enc_dict: dict[str, Any] = chardet.detect(raw_data) # type: ignore
-            encoding: str = enc_dict["encoding"] or "utf-8" # type: ignore
-            if enc_dict["confidence"] < 0.90:
-                encoding = "utf-8"
-            try:
-                content = raw_data.decode(encoding)  # type: ignore
-            except UnicodeDecodeError as _:
-                content = raw_data.decode(enc_dict["encoding"] or "utf-8") # type: ignore
-            return content.replace('\r\n', '\n')
+
+    def load(file_path: str) -> str:
+        try:
+            with open(file_path, "r", encoding="utf-8") as f:
+                return f.read().replace("\r\n", "\n")
+        except UnicodeDecodeError:
+            with open(file_path, "r", encoding="latin-1") as f:
+                return f.read().replace("\r\n", "\n")
 
     @staticmethod
     def save(file_path: str, content: str):
@@ -29,13 +21,3 @@ class Decoder:
             os.makedirs(base_path)
         with open(file_path, "w", encoding="utf-8", newline='\n') as file:
             file.write(content)
-
-    @staticmethod
-    def get_encoding(file_path: str) -> str:
-        with open(file_path, "rb") as file:
-            raw_data = file.read()
-            enc_dict = chardet.detect(raw_data)
-            encoding = enc_dict["encoding"]
-            if enc_dict["confidence"] < 0.90:
-                encoding = "utf-8"
-            return encoding or "utf-8"